Changelist: -modified two module tests to better fit requirements -changed behaviour for PUT on /preferences to also add the entry into the database if it doesn't exist already

+++ src/main/java/com/backend/fakedb/services/UserPreferencesService.java
@@ -88,7 +88,10 @@ public class UserPreferencesService {
         }
 
         var preferencesID = new UserPreferencesPK(user.get(), provider.get());
-        upRepository.deleteById(preferencesID);
+
+        if (upRepository.findById(preferencesID).isPresent()) {
+            upRepository.deleteById(preferencesID);
+        }
         upRepository.save(new UserPreferencesEntity(preferencesID, status));
     }
 }
